// src-ui/core/ipc/generated/domain/AdvancedConfig.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BypassConfig } from "./BypassConfig";
import type { LogLevel } from "./LogLevel";
import type { O3HookMode } from "./O3HookMode";

export type AdvancedConfig = { autoStart: boolean, offlineNotice: boolean, parseMultMsg: boolean, packetServer: string, packetBackend: string, enableLocalFile2Url: boolean, fileLog: boolean, consoleLog: boolean, fileLogLevel: LogLevel, consoleLogLevel: LogLevel, o3HookMode: O3HookMode, bypass: BypassConfig, };